add native filter helper
